Refactor and add tests.
